add detail activity for stock history
Added StockDetailActivity and activity_stock_detail files, modified the manifest, and modified MainActivity to open the new activity when a stock is clicked.

Currently the cursor data is being passed to the second activity through an intent.
All data is being extracted from the cursor as Strings and passed through the intent as Objects.
This could be improved